+++ data-final-fixes.lua
@@ -1,6 +1,6 @@
 require("__LarkenxK2SETweaks__/data/K2SE_logistic_pipes_data-final-fixes")
 require("__LarkenxK2SETweaks__/data/SE_buildings_in-space_data-final-fixes")
-require("__LarkenxK2SETweaks__/data/K2SE_buildings_lab.data-final-fixes")
+require("__LarkenxK2SETweaks__/data/K2SE_buildings_lab_data-final-fixes")
 require("__LarkenxK2SETweaks__/data/SE_buildings_on-spaceship_data-final-fixes")
-require("__LarkenxK2SETweaks__/data/K2SE_matterconversion_cubes.data-final-fixes")
-require("__LarkenxK2SETweaks__/data/K2SE_matterconversion_change_rates.data-final-fixes")
+require("__LarkenxK2SETweaks__/data/K2SE_matterconversion_cubes_data-final-fixes")
+require("__LarkenxK2SETweaks__/data/K2SE_matterconversion_change-rates_data-final-fixes")
